分类管理代码实现(完善删除逻辑)

+++ sky-server/src/main/java/com/sky/service/impl/categoryServicesImpl.java
@@ -2,11 +2,15 @@ package com.sky.service.impl;
 
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
+import com.sky.constant.MessageConstant;
 import com.sky.context.BaseContext;
 import com.sky.dto.CategoryDTO;
 import com.sky.dto.CategoryPageQueryDTO;
 import com.sky.entity.Category;
+import com.sky.exception.DeletionNotAllowedException;
 import com.sky.mapper.categoryMapper;
+import com.sky.mapper.dishMapper;
+import com.sky.mapper.setMealMapper;
 import com.sky.result.PageResult;
 import com.sky.service.categoryService;
 import org.springframework.beans.BeanUtils;
@@ -16,13 +20,19 @@ import org.springframework.stereotype.Service;
 import java.time.LocalDateTime;
 import java.util.List;
 
-import static com.sky.constant.StatusConstant.ENABLE;
+import static com.sky.constant.StatusConstant.DISABLE;
 
 @Service
 public class categoryServicesImpl implements categoryService {
 
     @Autowired
     categoryMapper categoryMapper;
+
+    @Autowired
+    dishMapper dishMapper;
+
+    @Autowired
+    setMealMapper setMealMapper;
 
     //分页查询菜品
     public PageResult page(CategoryPageQueryDTO categoryPageQueryDTO){
@@ -43,7 +53,7 @@ public class categoryServicesImpl implements categoryService {
         BeanUtils.copyProperties(categoryDTO,category);
 
         //补充菜品信息
-        category.setStatus(ENABLE);
+        category.setStatus(DISABLE);
         category.setCreateTime(LocalDateTime.now());
         category.setUpdateTime(LocalDateTime.now());
         category.setCreateUser(BaseContext.getCurrentId());
@@ -64,6 +74,19 @@ public class categoryServicesImpl implements categoryService {
 
     //删除对应id菜品分类
     public void delete(Integer id){
+
+        //查询当前分类是否关联菜品，如果关联则抛出异常
+        Integer count = dishMapper.CountById(id);
+        if(count>0){
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_DISH);
+        }
+
+        //查询当前分类是否关联套餐
+        count = setMealMapper.CountById(id);
+        if(count>0){
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_SETMEAL);
+        }
+        //删除分类
         categoryMapper.delete(id);
     }
